Fixed another compiler warning - hope I didn't break anything.
git-svn-id: svn+ssh://svn.code.sf.net/p/supertuxkart/code/main/trunk@12971 178a84e3-b1eb-0310-8ba1-8eac791a3b58

+++ src/guiengine/widgets/rating_bar_widget.hpp
@@ -51,7 +51,7 @@ namespace GUIEngine
         void setRating(float rating) { m_rating = rating; };
         
         /** Get the current value of the widget. */
-        int getRating() {return m_rating; };
+        float getRating() {return m_rating; };
         
         /** Change the number of star of the widget. */
         void setStarNumber(int star_number) { m_star_number = star_number; };
